requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.8
python-dotenv==1.2.1
Pillow==12.1.1
mServices>=1.0.22
